move styled components building to sep file with new optimization dirs
